move query integration into a separate package
also renamed init function
