+++ package.json
@@ -40,6 +40,7 @@
     "@playwright/test": "^1.52.0",
     "@tanstack/config": "^0.16.1",
     "@tanstack/react-query": "5.66.0",
+    "@tanstack/query-core": "5.66.0",
     "@types/node": "^22.10.2",
     "@types/react": "^19.0.8",
     "@types/react-dom": "^19.0.3",
@@ -77,7 +78,8 @@
       "eslint": "$eslint",
       "vite": "$vite",
       "@playwright/test": "$@playwright/test",
-      "@tanstack/react-query": "5.66.0",
+      "@tanstack/react-query": "$@tanstack/react-query",
+      "@tanstack/query-core": "$@tanstack/query-core",
       "@tanstack/history": "workspace:*",
       "@tanstack/router-core": "workspace:*",
       "@tanstack/react-router": "workspace:*",
@@ -90,6 +92,7 @@
       "@tanstack/virtual-file-routes": "workspace:*",
       "@tanstack/router-plugin": "workspace:*",
       "@tanstack/router-vite-plugin": "workspace:*",
+      "@tanstack/router-ssr-query-core": "workspace:*",
       "@tanstack/react-router-ssr-query": "workspace:*",
       "@tanstack/zod-adapter": "workspace:*",
       "@tanstack/valibot-adapter": "workspace:*",
