+++ packages/router-ssr-query-core/eslint.config.js
@@ -1,0 +1,14 @@
+// @ts-check
+
+import rootConfig from '../../eslint.config.js'
+
+export default [
+  ...rootConfig,
+  {
+    files: ['src/**/*.{ts,tsx}', 'tests/**/*.{ts,tsx}'],
+    rules: {
+      'unused-imports/no-unused-vars': 'off',
+      '@typescript-eslint/no-unnecessary-condition': 'off',
+    },
+  },
+]
