+++ packages/router-ssr-query-core/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src", "vite.config.ts", "tests", "vite-minify-plugin.ts"]
+}
